+++ chaosreport/__init__.py
@@ -23,7 +23,7 @@ import matplotlib.dates as mdates
 import maya
 from natural import date
 import pygal
-from pygal.style import DefaultStyle
+from pygal.style import DefaultStyle, LightColorizedStyle
 import pypandoc
 import semver
 
@@ -47,6 +47,8 @@ def generate_report_header(journal_paths: List[str],
     header_info["export_format"] = export_format
     header_info["tags"] = []
 
+    contribution_labels = []
+    contributions_by_level = []
     experiment_titles = []
 
     for journal_path in journal_paths:
@@ -55,8 +57,81 @@ def generate_report_header(journal_paths: List[str],
 
         experiment = journal.get("experiment")
         header_info["tags"].extend(experiment.get("tags", []))
+        contribs = experiment.get("contributions")
 
+        if not contribs:
+            continue
+
+        title = experiment["title"]
+        experiment_titles.append(title)
+        for contrib in contribs:
+            contribution_labels.append(contrib)
+            level = contribs[contrib]
+            contributions_by_level.append((title, level, contrib))
+
+    number_of_contributions = len(set(contribution_labels))
+    unique_contributions = sorted(set(contribution_labels))
+    header_info["num_experiments"] = len(experiment_titles)
+    header_info["num_distinct_contributions"] = number_of_contributions
     header_info["tags"] = set(header_info["tags"])
+
+    dist_chart = pygal.Bar(
+        print_values=True, print_values_position='top', show_legend=False,
+        show_y_labels=False, legend_at_bottom=True)
+    dist_chart.title = 'Organization Contributions Distribution'
+    dist_chart.x_labels = unique_contributions
+    dist_chart.add(
+        "", [
+            contribution_labels.count(contrib)
+            for contrib in unique_contributions
+        ])
+
+    if export_format in ["html", "html5"]:
+        header_info["contribution_distribution"] = dist_chart.render(
+                disable_xml_declaration=True)
+    else:
+        header_info["contribution_distribution"] = b64encode(
+            cairosvg.svg2png(
+                bytestring=dist_chart.render(), dpi=72)).decode("utf-8")
+
+    contributions = {}
+    for title in experiment_titles:
+        contributions[title] = [None] * number_of_contributions
+
+    contribution_labels = list(unique_contributions)
+    for (title, level, contrib) in contributions_by_level:
+        idx = contribution_labels.index(contrib)
+        amount = 0
+        if level == "high":
+            amount = 0.75
+        elif level == "medium":
+            amount = 0.50
+        elif level == "low":
+            amount = 0.25
+        elif level == "none":
+            amount = -0.1
+        else:
+            continue
+        contributions[title][idx] = amount
+
+    chart = pygal.Dot(
+        legend_at_bottom_columns=1, show_y_labels=False, legend_at_bottom=True,
+        show_legend=True, x_label_rotation=30, style=LightColorizedStyle,
+        interpolate='hermite')
+    chart.title = 'Organization Contributions Impact'
+    chart.x_labels = contribution_labels
+    for title in contributions:
+        chart.add(
+            title, contributions[title], fill=False, allow_interruptions=True)
+
+    if export_format in ["html", "html5"]:
+        header_info["contributions"] = chart.render(
+                disable_xml_declaration=True)
+    else:
+        header_info["contributions"] = b64encode(
+            cairosvg.svg2png(
+                bytestring=chart.render(), dpi=72)).decode("utf-8")
+
     header = header_template.render(header_info)
     return header
 
@@ -84,7 +159,7 @@ def generate_report(journal_path: str, export_format: str = "markdown") -> str:
     journal["today"] = datetime.now().strftime("%d %B %Y")
 
     generate_chart_from_metric_probes(journal, export_format)
-    add_contribution_model(experiment)
+    add_contribution_model(journal, export_format)
     template = get_report_template(journal["chaoslib-version"])
     report = template.render(journal)
 
@@ -394,15 +469,32 @@ def generate_from_vegeta_result(run: Run, export_format: str):
         add_chart(status_distribution())
 
 
-def add_contribution_model(experiment: Experiment):
+def add_contribution_model(journal: Journal, export_format: str):
     """
     Expose the contribution of that experiment to the report.
-
-    As this is part of an extension, we bubble it up to the experiment itself
-    for rendering purpose.
     """
-    for extension in experiment.get("extensions", []):
-        contributions = extension.get("contributions")
-        if contributions:
-            experiment["contributions"] = contributions
-            break
+    experiment = journal.get("experiment")
+    contributions = experiment.get("contributions")
+    if not contributions:
+        return
+
+    chart = pygal.HorizontalBar()
+    chart.title = 'Organization Contributions Impact'
+    for (contribution, impact) in contributions.items():
+        value = 0
+        if impact == "high":
+            value = 0.75
+        elif impact == "medium":
+            value = 0.5
+        if impact == "low":
+            value = 0.25
+
+        chart.add(contribution, value)
+
+    if export_format in ["html", "html5"]:
+        experiment["contributions_chart"] = chart.render(
+            disable_xml_declaration=True)
+    else:
+        experiment["contributions_chart"] = b64encode(
+            cairosvg.svg2png(
+                bytestring=chart.render(), dpi=72)).decode("utf-8")